requirements.txt
openai
faiss-cpu
scikit-learn
python-pptx
numpy
sentence-transformers
streamlit
groq